Changes to Post and Get in PHP
The PHP Should now output all of the content and the title of the post during a GET. During the POST there is a minor issue of creating a button ID for the reply buttons associated with comments, the plan I had in mind was to assign values to the buttons like I did with the standard submit comment button. That way I would know if the user was submitting a comment or a reply and change the table name to match.